+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/ui/viewmodel/MainViewModel.kt
@@ -146,6 +146,9 @@ class MainViewModel(
             is NetworkTestAction.PerformOkHttpRequest -> {
                 networkTestingRepository.performOkHttpRequest()
             }
+            is NetworkTestAction.PerformOkHttpFailureBeforeResponseHeaders -> {
+                networkTestingRepository.performOkHttpFailureBeforeResponseHeaders()
+            }
             is NetworkTestAction.PerformGraphQlRequest -> {
                 networkTestingRepository.performGraphQlRequest()
             }