+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 **Fixed**
 
 - SDK was erronously reporting `_jvm_used_percent` in Resource Utilization logs using the device Locale
+- Fixed OkHttp logs including an invalid `_protocol` value when a request fails before receiving a response.
 
 ### iOS